Collision somewhat fixed
still not quite right though.  The bounding spheres appear to be not at
all related to where the models get drawn.  I hve a class with helper
methods to draw the bounding spheres, but that's not quite right either
as it doesn't draw them to scale or in the right position either.  so
all in all, al ot of not working stuff...