More updates
Goal Model ingame, it's hard to see ATM, but it randomly spawns and will
respawn if it spawns inside of an asteroid.  It also respawn when you
collide with it.  Timer Added, currently 30 second games but you can
change that easily in Game1 public static's you can also adjust how much
time is added when you successfully hit a goal.